+++ modules/focus-navigation/src/__tests__/EngineInterface.spec.lua
@@ -6,7 +6,9 @@ local JestGlobals = require(Packages.Dev.JestGlobals)
 local EventPropagation = require(Packages.EventPropagation)
 
 local it = JestGlobals.it
+local describe = JestGlobals.describe
 local expect = JestGlobals.expect
+local beforeEach = JestGlobals.beforeEach
 local afterEach = JestGlobals.afterEach
 
 local createGuiObjectTree = require(script.Parent.createGuiObjectTree)
@@ -68,8 +70,45 @@ end)
 
 it("should be able to focus child of non-Selectable Player descendant", function()
 	mountedTree = getMountedGui(PlayerGui)
-	service = FocusNavigationService.new(EngineInterface.CoreGui)
+	service = FocusNavigationService.new(EngineInterface.PlayerGui)
 	service:focusGuiObject(mountedTree.container, false)
 	expect(GuiService.SelectedObject).toEqual(mountedTree.button)
 	expect(GuiService.SelectedCoreObject).toBeNil()
+end)
+
+describe("warnings", function()
+	local coreTarget, playerTarget
+	beforeEach(function()
+		coreTarget = Instance.new("TextButton")
+		coreTarget.Parent = game:GetService("CoreGui")
+
+		playerTarget = Instance.new("TextButton")
+		playerTarget.Parent = (game:GetService("Players").LocalPlayer :: any).PlayerGui
+	end)
+	afterEach(function()
+		coreTarget:Destroy()
+		playerTarget:Destroy()
+	end)
+
+	it("warns when trying to focus outside of target gui interface", function()
+		expect(function()
+			EngineInterface.CoreGui.setSelection(coreTarget)
+		end).toWarnDev({})
+		expect(function()
+			EngineInterface.CoreGui.setSelection(playerTarget)
+		end).toWarnDev({
+			"Attempting to focus non%-CoreGui descendant 'Players.DummyTestUser.PlayerGui.TextButton' "
+				.. "from a CoreGui FocusNavigationService ",
+		})
+
+		expect(function()
+			EngineInterface.PlayerGui.setSelection(playerTarget)
+		end).toWarnDev({})
+		expect(function()
+			EngineInterface.PlayerGui.setSelection(coreTarget)
+		end).toWarnDev({
+			"Attempting to focus CoreGui descendant 'CoreGui.TextButton' "
+				.. "from a non%-CoreGui FocusNavigationService ",
+		})
+	end)
 end)