+++ modules/utils/src/init.lua
@@ -2,6 +2,7 @@
 local createSignal = require(script.createSignal)
 local shallowEqual = require(script.shallowEqual)
 local getAncestors = require(script.getAncestors)
+local mockableWarn = require(script.mockableWarn)
 
 export type Signal<T> = createSignal.Signal<T>
 export type FireSignal<T> = createSignal.FireSignal<T>
@@ -11,4 +12,5 @@ return {
 	createSignal = createSignal,
 	shallowEqual = shallowEqual,
 	getAncestors = getAncestors,
+	mockableWarn = mockableWarn,
 }